devices: Move i2c code into a submodule
Adding support for additional device types will thus be a matter of
adding sibling submodules.
