+++ src/main.rs
@@ -45,6 +45,7 @@ use sensor::{
 	SensorIntfData,
 	SensorMap,
 };
+use devices::i2c::I2CDeviceMap;
 
 /// The dbus name claimed by the daemon.
 const DBUS_NAME: &str = "xyz.openbmc_project.OmniSensor";
@@ -134,7 +135,7 @@ pub struct DaemonState {
 	/// All extant sensors, by name (active and inactive alike).
 	sensors: Mutex<SensorMap>,
 	/// All managed (dynamic) I2C devices.
-	i2cdevs: Mutex<devices::I2CDeviceMap>,
+	i2cdevs: Mutex<I2CDeviceMap>,
 	/// Our dbus connection.
 	bus: Arc<SyncConnection>,
 	/// dbus object server registry...thing.
@@ -225,7 +226,7 @@ async fn main() -> ErrResult<()> {
 	let daemonstate = DaemonState {
 		config: Mutex::new(cfg),
 		sensors: Mutex::new(SensorMap::new()),
-		i2cdevs: Mutex::new(devices::I2CDeviceMap::new()),
+		i2cdevs: Mutex::new(I2CDeviceMap::new()),
 		bus,
 		crossroads: SyncMutex::new(cr),
 		sensor_intfs,
